Massive code cleanup removing most GCC warnings, some potential problems and a few minor bugs.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@444 01821946-851b-0410-b7f4-f1477c70e57a

+++ arch/linux/ukali.c
@@ -64,11 +64,11 @@ int knix_Send(int hand, char *data, int len) {
 
 int knix_Recv(int hand, char *data, int len) {
 	struct sockaddr_in taddr;
-	int tlen;
+	unsigned int tlen;
 
 	tlen = sizeof(taddr);
 
-	return recvfrom(hand, data, len, 0, (struct sockaddr *)&taddr, &tlen);
+	return (int) recvfrom(hand, data, len, 0, (struct sockaddr *)&taddr, &tlen);
 
 }
